Perform S3 directory deletion with batch requests
Delete all the objects which are found under the specified path
prefix in batches in order to improve the performance of
the operation.

This operation is customised for the needs of Trino and does
intentionally not intend to add unnecessary complexity
for dealing with the various quirks of s3 compatible object
storage systems that are not relevant for the use cases of Trino.